Small bug fixes

+++ chef/run.sh
@@ -2,25 +2,29 @@
 
 echo 
 
-echo '*********************'
-echo '* VENDOR COOKBOOKS'
-echo '*********************'
+if where -q berks; then
+	echo '*********************'
+	echo '* VENDOR COOKBOOKS'
+	echo '*********************'
 
-pushd cookbooks/mybook 
-rm -Rf  ../vendored/cookbooks/mybook 
-berks vendor ../vendored/cookbooks 
-rm -Rf ../vendored/cookbooks/mybook
-popd
+	pushd cookbooks/mybook 
+	rm -Rf  ../vendored/cookbooks/mybook 
+	berks vendor ../vendored/cookbooks 
+	rm -Rf ../vendored/cookbooks/mybook
+	popd
 
-echo
+	echo
+fi
 
 
-echo
-echo '*********************'
-echo '* FOODCRITIC MYBOOK'
-echo '*********************'
+if where -q foodcritic; then
+	echo
+	echo '*********************'
+	echo '* FOODCRITIC MYBOOK'
+	echo '*********************'
 
-foodcritic cookbooks/mybook
+	foodcritic cookbooks/mybook
+fi
 
 #:: runs a cookbok in solo mode in the current path
 
@@ -29,5 +33,5 @@ echo '*********************'
 echo '* RUN CLIENT'
 echo '*********************'
 
-
+rm -Rf local_mode_cache
 chef-client --local-mode --minimal-ohai --config client.rb "$@"
